コンポーネントをファイルに切り出す(一部変更追加)
