+++ src/App.js
@@ -1,59 +1,8 @@
 import React, { useRef } from "react";
 import { useTodo } from "./hooks/useTodo";
-
-const TodoTitle = ({ title, as }) => {
-  if (as === "h1") {
-    return <h1>{title}</h1>;
-  } else if (as === "h2") {
-    return <h2>{title}</h2>;
-  } else {
-    return <p>{title}</p>;
-  }
-};
-
-const TodoList = ({
-  todoList,
-  toggleTodoListItemStatus,
-  deleteTodoListItem,
-}) => {
-  return (
-    <ul>
-      {todoList.map((todo) => (
-        <TodoItem
-          todo={todo}
-          key={todo.id}
-          toggleTodoListItemStatus={toggleTodoListItemStatus}
-          deleteTodoListItem={deleteTodoListItem}
-        />
-      ))}
-    </ul>
-  );
-};
-
-const TodoItem = ({ todo, toggleTodoListItemStatus, deleteTodoListItem }) => {
-  const handleToggleTodoListItemStatus = () =>
-    toggleTodoListItemStatus(todo.id, todo.done);
-  const handleDeleteTodoListItem = () => deleteTodoListItem(todo.id);
-
-  return (
-    <li>
-      {todo.content}
-      <button onClick={handleToggleTodoListItemStatus}>
-        {todo.done ? "完了リストへ" : "未完了リストへ"}
-      </button>
-      <button onClick={handleDeleteTodoListItem}>削除</button>
-    </li>
-  );
-};
-
-const TodoAdd = ({ inputEl, handleAddTodoListItem }) => {
-  return (
-    <>
-      <textarea ref={inputEl} />
-      <button onClick={handleAddTodoListItem}>+ TODOを追加</button>
-    </>
-  );
-};
+import { TodoTitle } from "./components/TodoTitle";
+import { TodoList } from "./components/TodoList";
+import { TodoAdd } from "./components/TodoAdd";
 
 function App() {
   const {
@@ -83,18 +32,20 @@ function App() {
         handleAddTodoListItem={handleAddTodoListItem}
       />
 
-      <TodoTitle title="未完了TODOリスト" as="h2" />
       <TodoList
         todoList={inCompletedList}
         toggleTodoListItemStatus={toggleTodoListItemStatus}
         deleteTodoListItem={deleteTodoListItem}
+        title="未完了TODOリスト"
+        as="h2"
       />
 
-      <TodoTitle title="完了TODOリスト" as="h2" />
       <TodoList
         todoList={completedList}
         toggleTodoListItemStatus={toggleTodoListItemStatus}
         deleteTodoListItem={deleteTodoListItem}
+        title="完了TODOリスト"
+        as="h2"
       />
     </>
   );
