+++ src/components/TodoAdd.jsx
@@ -1,0 +1,10 @@
+const TodoAdd = ({ inputEl, handleAddTodoListItem }) => {
+  return (
+    <>
+      <textarea ref={inputEl} />
+      <button onClick={handleAddTodoListItem}>+ TODOを追加</button>
+    </>
+  );
+};
+
+export { TodoAdd };
